refactor: move isPlainObject and mergeDeep to @internals/utils
Agent-Logs-Url: https://github.com/kubb-labs/kubb/sessions/91a079f9-3de3-44e3-9e52-7aac28a2ae95
