+++ internals/utils/src/index.ts
@@ -10,7 +10,7 @@ export { canUseTTY, isCIEnvironment, isGitHubActions } from './env.ts'
 export { BuildError, getErrorMessage, toCause, toError } from './errors.ts'
 export { clean, exists, findPackageJSON, getRelativePath, read, readSync, write } from './fs.ts'
 export { executeIfOnline } from './network.ts'
-export { serializePluginOptions } from './object.ts'
+export { isPlainObject, mergeDeep, serializePluginOptions } from './object.ts'
 export { onProcessExit } from './onProcessExit.ts'
 export type { PackageManagerInfo, PackageManagerName } from './packageManager.ts'
 export { detectPackageManager } from './packageManager.ts'
