+++ packages/adapter-oas/src/factory.ts
@@ -1,5 +1,5 @@
 import path from 'node:path'
-import { URLPath } from '@internals/utils'
+import { URLPath, mergeDeep } from '@internals/utils'
 import type { AdapterSource } from '@kubb/core'
 import { bundle, loadConfig } from '@redocly/openapi-core'
 import OASNormalize from 'oas-normalize'
@@ -7,19 +7,6 @@ import swagger2openapi from 'swagger2openapi'
 import { MERGE_DEFAULT_TITLE, MERGE_DEFAULT_VERSION, MERGE_OPENAPI_VERSION } from './constants.ts'
 import { isOpenApiV2Document } from './guards.ts'
 import type { Document } from './types.ts'
-
-function mergeDeep(target: Record<string, unknown>, source: Record<string, unknown>): Record<string, unknown> {
-  const result: Record<string, unknown> = { ...target }
-  for (const key of Object.keys(source)) {
-    const sv = source[key]
-    const tv = result[key]
-    result[key] =
-      sv !== null && typeof sv === 'object' && !Array.isArray(sv) && tv !== null && typeof tv === 'object' && !Array.isArray(tv)
-        ? mergeDeep(tv as Record<string, unknown>, sv as Record<string, unknown>)
-        : sv
-  }
-  return result
-}
 
 export type ParseOptions = {
   canBundle?: boolean
@@ -71,7 +58,7 @@ export async function parseDocument(pathOrApi: string | Document, { canBundle = 
 /**
  * Deep-merges multiple OpenAPI documents into a single `Document`.
  *
- * Each document is parsed independently then recursively merged with `remeda`'s `mergeDeep`.
+ * Each document is parsed independently then recursively merged.
  * Throws when the input array is empty.
  *
  * @example
